dt-bindings: net: ethernet-switch: Rename $defs "base" to 'ethernet-ports'
commit b9823df7bbad42bffa1d13e65c11c1e3ae09069c upstream.

The name "base" is misleading as the definition is for a complete schema
definition without additional properties allowed, not a "base class".
Align the same to be the same as dsa.yaml. This schema file without any
json pointer path is the base schema which can be extended.

There are not yet any references to $defs/base to update.

Signed-off-by: Rob Herring <robh@kernel.org>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Acked-by: Arınç ÜNAL <arinc.unal@arinc9.com>
Acked-by: Linus Walleij <linus.walleij@linaro.org>
Link: https://lore.kernel.org/r/20231016-dt-net-cleanups-v1-5-a525a090b444@kernel.org
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>