net: ethernet: am65-cpsw: Set default TX channels to maximum
commit be397ea3473d24ce596f3aaa22a55af9b004fed8 upstream.

am65-cpsw supports 8 TX hardware queues. Set this as default.

The rationale is that some am65-cpsw devices can have up to 4 ethernet
ports. If the number of TX channels have to be changed then all
interfaces have to be brought down and up as the old default of 1
TX channel is too restrictive for any mqprio/taprio usage.

Another reason for this change is to allow testing using
kselftest:net/forwarding:ethtool_mm.sh out of the box.

Signed-off-by: Roger Quadros <rogerq@kernel.org>
Reviewed-by: Andrew Lunn <andrew@lunn.ch>
Link: https://lore.kernel.org/r/20231117121755.104547-4-rogerq@kernel.org
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>