arm64: dts: ti: k3-am64: Enable SDHCI nodes at the board level
commit 3b6345e3fcf4c93a79f396121cd0e6f98f04da13 upstream.

SDHCI nodes defined in the top-level AM64 SoC dtsi files are incomplete
and will not be functional unless they are extended.

As the attached SD/eMMC is only known about at the board integration level,
these nodes should only be enabled when provided with this information.

Disable the SDHCI nodes in the dtsi files and only enable the ones that
are actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20231117163339.89952-2-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>