ASoC: tas2562: move tas2563 from tas2562 driver to tas2781 driver
commit 645994d21287a1ad2f637818d737f7a3d84e97d7 upstream.

Move tas2563 from tas2562 driver to tas2781 driver to unbind tas2563 from
tas2562 driver code and bind it to tas2781 driver code, because tas2563
only work in bypass-DSP mode with tas2562 driver. In order to enable DSP
mode for tas2563, it has been moved to tas2781 driver. As to the hardware
part, such as register setting and DSP firmware, all these are stored in
the binary firmware. What tas2781 drivder does is to parse the firmware
and download it to the chip, then power on the chip. So, tas2781 driver
can be resued as tas2563 driver. Only attention will be paid to
downloading corresponding firmware.

Signed-off-by: Shenghao Ding <shenghao-ding@ti.com>
Link: https://lore.kernel.org/r/20240104145721.1398-2-shenghao-ding@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>